Power Up
I made a prefab for a slow down power up. There is a base class for all powerups that just hold duration and an isActive property to use to start a timer. So in new inherited classes should use base.Update() to start the timer and then do their things in an if statment that checks isActive. Slow down power up now works perfectly on the android device. Also changed a few settings so things are functional on android.