Fixing Bugs
Changed some values and objects now spawn based on a random number from 0-100 so each thing has a certain percentage of spawing so we dont get a million power ups. Values still need to change. Main mechanic doesnt work for gravity switching sometimes and certain values need to be hardcoded so things dont spawn in the middle of the game screen